+++ packages/frontend/web-app/package.json
@@ -15,6 +15,8 @@
     "eject": "react-scripts eject"
   },
   "dependencies": {
+    "@material-ui/core": "^4.11.0",
+    "@material-ui/styles": "^4.10.0",
     "@packages/backend-ms-gateway": "*",
     "react": "^16.14.0",
     "react-dom": "^16.14.0"
