--- packages/frontend/web-app/src/App.tsx
@@ -1,32 +1,0 @@
-import './App.css';
-
-import * as ApiTypes from '@packages/backend-ms-gateway/build/src/types/User';
-import React from 'react';
-
-import logo from './logo.svg';
-
-function App() {
-  const user: ApiTypes.User = { name: 'Bob' };
-
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <p>The user is named {user.name}</p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
-}
-
-export default App;
