+++ packages/frontend/web-app/src/index.tsx
@@ -1,9 +1,11 @@
+/* istanbul ignore file */
+
 import './index.css';
 
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './App';
+import App from './app/App';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
